get player stat avgs

+++ sbs-v1-rust-backend/rust_backend/src/db/nba_games_historical_mongo_dao.rs
@@ -4,7 +4,7 @@ use mongodb::bson::{doc, Document, from_document};
 use crate::models::db::nba_games_historical::NbaGamesHistorical;
 use mongodb::error::Result;
 
-pub async fn get_nba_games_by_team_and_season(collection: &Collection<Document>, teamNickname: &str, season: f32) -> Result<Vec<NbaGamesHistorical>> {
+pub async fn get_nba_games_by_team_and_season(collection: &Collection<Document>, teamNickname: &str, season: f64) -> Result<Vec<NbaGamesHistorical>> {
    let query = doc! { 
         "$or": [ 
             { "teamsHomeNickname": teamNickname },
