+++ sbs-v1-rust-backend/rust_backend/src/models/db/nba_odds_historical.rs
@@ -35,7 +35,7 @@ pub struct Outcome {
     name: String,
     price: i32,
     description: Option<String>,
-    point: Option<f32>
+    point: Option<f64>
 }
