+++ sbs-v1-rust-backend/rust_backend/src/models/services/get_nba_odds_by_team_and_season_request.rs
@@ -3,5 +3,5 @@ use serde::Deserialize;
 #[derive(Debug, Deserialize)]
 pub struct GetNbaOddsByTeamAndSeasonRequest {
     pub teamName: String,
-    pub season: f32,
+    pub season: f64,
 }
